- Reorganize Foundation.Base - Proxy for static Sitecore.Context
(cherry picked from commit ceefb853762642eae5cbdf69c88362d18454d92f)
