--- src/Foundation/Base/code/Services/Mapping/IMapService.cs
@@ -1,41 +1,0 @@
-﻿//    Copyright 2020 EPAM Systems, Inc.
-// 
-//    Licensed under the Apache License, Version 2.0 (the "License");
-//    you may not use this file except in compliance with the License.
-//    You may obtain a copy of the License at
-// 
-//      http://www.apache.org/licenses/LICENSE-2.0
-// 
-//    Unless required by applicable law or agreed to in writing, software
-//    distributed under the License is distributed on an "AS IS" BASIS,
-//    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//    See the License for the specific language governing permissions and
-//    limitations under the License.
-
-namespace Wooli.Foundation.Base.Services.Mapping
-{
-    /// <summary>
-    /// Maps models using AutoMapper
-    /// </summary>
-    public interface IMapService
-    {
-        /// <summary>
-        /// Wraps AutoMapper.Mapper.Map
-        /// </summary>
-        /// <typeparam name="TSource">Type of source model</typeparam>
-        /// <typeparam name="TDestination">Type of destination model</typeparam>
-        /// <param name="source">Source model</param>
-        /// <returns>Mapped model</returns>
-        TDestination Map<TSource, TDestination>(TSource source);
-
-        /// <summary>
-        /// Wraps AutoMapper.Mapper.Map
-        /// </summary>
-        /// <typeparam name="TSource">Type of source model</typeparam>
-        /// <typeparam name="TDestination">Type of destination model</typeparam>
-        /// <param name="source">Source model</param>
-        /// <param name="destination">Destination model</param>
-        /// <returns>Mapped model</returns>
-        TDestination Map<TSource, TDestination>(TSource source, TDestination destination);
-    }
-}
