+++ src/Foundation/Base/code/Wooli.Foundation.Base.csproj
@@ -435,17 +435,17 @@
     <Compile Include="Models\Logging\Log.cs" />
     <Compile Include="Models\Logging\CommonLog.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Services\Configuration\ConfigurationProvider.cs" />
+    <Compile Include="Providers\Configuration\ConfigurationProvider.cs" />
+    <Compile Include="Providers\Configuration\IConfigurationProvider.cs" />
+    <Compile Include="Context\ISitecoreContext.cs" />
+    <Compile Include="Context\SitecoreContext.cs" />
     <Compile Include="Services\Configuration\ConfigurationService.cs" />
-    <Compile Include="Services\Configuration\IConfigurationProvider.cs" />
     <Compile Include="Services\Configuration\IConfigurationService.cs" />
     <Compile Include="Services\Context\ISitecoreContext.cs" />
     <Compile Include="Services\Context\SitecoreContext.cs" />
     <Compile Include="Services\Logging\CommonLogService.cs" />
     <Compile Include="Services\Logging\ILogService.cs" />
     <Compile Include="Services\Logging\LogService.cs" />
-    <Compile Include="Services\Mapping\IMapService.cs" />
-    <Compile Include="Services\Mapping\MapService.cs" />
     <Compile Include="Services\Pipeline\IPipelineService.cs" />
     <Compile Include="Services\Pipeline\PipelineService.cs" />
     <Compile Include="Services\Tracking\TrackingService.cs" />
